requirements.txt
Flask==3.0.0
boto3==1.34.0
gunicorn==21.2.0
prometheus-flask-exporter==0.23.0
python-dotenv==1.0.0
flask-restx==1.3.0
pytest==7.4.3
pytest-cov==4.1.0
werkzeug==3.0.1
